`auto_config`: Inline calls to functions that return `Buildable`'s.
`auto_config` "inlines" calls to other `auto_config` functions.
But there are some config building patterns that can't be
expressed with `auto_config` (intentionally!). In order to allow
users to continue to use `auto_config` in a fully compositional
manner, we should equivalently treat functions that are annotated
to return `fdl.Buildable`'s.

PiperOrigin-RevId: 415915297