Support command line overrides containing collections.
Previously, you could set arbitrary arguments in a config on the command line
so long as the path contained _only_ `fdl.Buildable` types. This is suboptimal
because often config's contain collections. For example, a neural network model
contains a list of "modules" (or layers) within it. Before this change, you
could not set layer properties with command line overrides (e.g.
`--fdl.model.layers[3].dropout=False` would not work). With this change,
that all works now!

PiperOrigin-RevId: 427649243