Add `fdl.update_callable` for in-place updates.
Previously, it was not obvious how to swap highly-aliased `Buildable` for
one that would produce a different type. This sort of mutation could be handy
when swapping out a Dense layer for a Sparse (MoE) layer, or one optimizer
for another.

This change adds `fdl.update_callable` which will update the callable that
backs a `fdl.Buildable` in-place. Along the way, it will also verify that
previously set arguments continue to be valid. If there are invalid arguments,
a TypeError will be raised and include the names of _all_ invalid arguments for
the new callable.

A future commit could extend the `update_callable` API to automatically drop
new invalid params as required.

In order to support time-travel debugging for `update_callable` calls, a new
`__fn_or_cls__` entry has been added to `Buildable`'s `__argument_history__`.

Prior art: Lingvo HParams exposes the underlying callable as `params.cls`.

PiperOrigin-RevId: 427881750