Switch placeholders to `TaggedValue` and new class `Tag`s.
To arrive at this change, I prototyped a number of different design
alternatives, including decorators instead of metaclasses. However,
I settled on the metaclass implementation because it works with inheritance
between tags. Additionally, I explored using a typed `description: str`
field, instead of using the python docstring. While the typed property
approach did have (nicer?) autocomplete in IDEs, I found that after
defining a few `Tag`s, I'd be copy-pasting the same text from the
docstring to the description property. As a result, I thought it made
most sense to just unify them. In visualization tools, we should just
grab the first line if space is constrained (as the docstring convention
that the first line should be short & reasonable very much works in our
favor here). That said, should we decide to change the API to allow tags to
define a default value, we can use the same mechanism.

This change is the first of a sequence to potentially switch more of Fiddle's
mechanisms, including potentially a change to how collections are handled
with the goal of ensuring that placeholders/taggedvalues are a little more
reliable when used in practice.

PiperOrigin-RevId: 428114812