Add `__fiddle_init__ ` feature.
While Fiddle is designed around the principle that the configuration system
should only minimally intrude on the library, if a library would like to
customize itself for the configuration system, some usability wins are
possible.

This change allows callables (either functions or classes) to optionally define
a `__fiddle_init__ ` property that will be automatically called if present
immediately after `Buildable` creation.

PiperOrigin-RevId: 428951878