Improve subclass compatibility of __copy__/__deepcopy__ implementations.
It turns out that there were a number of Buildable subclasses floating around that didn't support __copy__ or __deepcopy__, since the signature of the constructor for those types was different from that of the base Buildable.

We could add support to them by directly implementing __copy__ and __deepcopy__ on those types... however, that feels a bit onerous and error prone in general (easy to forget, as is evidenced by the fact that it happened already for at least three types).

This CL implements another approach that I think will more frequently do the right thing and simplify subclass implementations. That said, it also has some drawbacks, since it goes behind the back of the subclass's __init__ method... curious to hear thoughts and opinions on the best approach to providing reliable __copy__/__deepcopy__ support.

PiperOrigin-RevId: 445802279