Add ability to register custom traversable nodes in daglish.
This also addresses the issue of copying certain Buildable subclasses like DictConfig and NamespaceConfig.

The overall design here is heavily based on the Jax tree_util approach.

Note that this supports custom registries for the register_node_traverser and find_node_traverser functions; this will be used in the upcoming serialization CL which does its own traversal and needs its own registry for custom serializable types. There isn't currently a mechanism to support using a custom registry for daglish traversal itself... I'm hesitant to add one, but if we find we need one we could do something like add a context manager to permit that.

PiperOrigin-RevId: 451041472