Allow tags to be set via dataclass metadata.
To reduce the need to reach for very powerful mechanisms (e.g.
`__fiddle_init__`), we enable select customizations to be enabled via metadata
attached to the `dataclasses.Field` type.

Adding tags is the first of a series of pieces of metadata we intend to place
here. Stay tuned for more in the future.

PiperOrigin-RevId: 462708269