Add `auto_config` support for `@classmethod`'s.
Previously, `auto_config` functions could only be used as top-level free
functions, or as `@staticmethod`'s. This is somewhat limiting. This change
introduces a new type (`AutoConfig`) that wraps the underlying function and its
`as_buildable` pair to customize the descriptor protocol implementation to bind
both the main method, and the `as_buildable` method.

PiperOrigin-RevId: 462729600