Automatically instantiate `auto_config`'d `default_factory`'s.
When a `dataclasses.field`'s `default_factory` is an `auto_config`'d function,
the Fiddle wrapper should automatically capture a reference to the
`as_buildable` function, and cause it to be automatically called when
instantiating a new `fdl.Buildable` for the dataclass itself.

Note: the design of the field metadata is intentionally generic such that other
behaviors (e.g. automatically instantiating an empty list when creating the
`fdl.Buildable`) can be supported from the same metadata structure. (It also
happens to simplify the dependency graph too.)

Finally, this is the first change to begin moving some internals of Fiddle to a
`_src` subdirectory (following JAX and other Python projects examples) to
provide a strong signal that certain bits of Fiddle should not be considered
part of the public API.

PiperOrigin-RevId: 468367675