Add `auto_unconfig` to plug in explicit config manipulation fns.
Currently, auto_config inspects the return annotation of functions to see if it
returns a `fdl.Buildable` type, and if so, auto-inlines the function. This is
a subtle and somewhat surprising behavior. In the process of making
`auto_config` more robust in preparation for moving out of `experimental`, this
sort of subtle behavior should be made explicit.

`auto_unconfig` is the dual to `auto_config`; `auto_config` takes a normal
Python function and creates a second `as_buildable` function that returns a
`fdl.Config` representing the function body, and `auto_unconfig` takes a
function that returns a `fdl.Config` and creates a second function that returns
a regular Python object.

For additional context, see the [previous commit][prev_commit] that modified
`auto_config`'s inline behavior.

[prev_commit]: https://github.com/google/fiddle/commit/58f01c69fccebeae682ccbf6ea5e759ea9ec44cd),

PiperOrigin-RevId: 469074731