`auto_config` functions should operate on a configurable policy.
Previously, `auto_config` used a couple heuristics to determine whether a call
should be left as-is, or if it should be converted into a `fdl.Config`
representation. While this worked for small examples, this does not work
reliably in more contexts. Instead, to ensure a consistent, reliable, and
flexible `auto_config` experience, we instead explicitly enumerate the set of
functions that should be evaluated and allow users to plug in their own
custom policy for inline evaluation.

The policy is intentionally customizable to allow users to add their own
domain-specific functions. One example might be a custom wrapper around
`typing.cast` to perform some additional validation.

This strategy is in contrast to attempting to restrict the use of control flow.
As a result, we hope to eventually enable control flow by default.

PiperOrigin-RevId: 469312791