Allow `raise` expressions inside `auto_config` functions.
The normal Python-path could reasonably want to validate preconditions or
verify other invariants. A natural way to express this is with exceptions.
Since `raise` has a reasonable interpretation for the initial evaluation of the
`auto_config` function, this change allows `raise` statements to be enabled
inside `auto_config` functions when control flow is enabled.

One key downside of this design is, the invariant checking only occurs during
the initial evaluation of the `auto_config` function. This likely doesn't occur
at `fdl.build` time, and thus, it's possible for some configurations to
partially bypass the checks.

PiperOrigin-RevId: 475046794