Require auto_config above classmethod/staticmethod to support Python 3.8/3.11+.
This CL makes auto_config descriptor forwarding work properly with classmethod/staticmethod objects, allowing auto_config to be placed above classmethod/staticmethod decorators, and has the nice side of effect of allowing the _BoundAutoConfig type to be removed entirely, which feels like a cleanup.

To enforce this ordering requirement, it implements best-effort check that examines the decorator list to determine if classmethod/staticmethod is present above auto_config. It also changes existing usage to match the new requirement.

Another potential mechanism to check this would be to unwrap the function and look at the name of the first argument, triggering if it is not self but the __qualname__ of the function suggests it is a method ("SomeClass.some_fn"...). This feels hackier, but would probably be able to catch cases where auto_config has been renamed or isn't applied as a decorator. Or maybe both approaches should be implemented for maximal coverage?

PiperOrigin-RevId: 490304872