Fix bug in `special_value_codegen.register_exact_value` that caused graphviz rendering to render integers as dtypes.
In particular, `_EXACT_VALUE_LOOKUP` is a table used to record repr strings for special values (such as `tf.float32`).  The intention was that it only be used for a value if that value was identical to a key (`key is val`); but the implementation was using it whenever the value was equal to a key (`key == val`).  And since `tf.float32 == 1`, this meant that we displayed `1` as `tf.float32`.

PiperOrigin-RevId: 492353728