Change the decorator API for defining default factories for functions.
Rather than passing the default factories to the decorator, we specify the default factory using the default value for each parameter.  This has a number of advantages:

* It keeps all the information about each parameter in one place.
* It makes pytype happier, since pytype knows that the argument has a default value.
* It more closely parallels the way dataclass default factories are defined.

One drawback of this change is that it introduces a new opportunity for user error -- namely, the user can set a default value with `arg_factory.default_factory`, but forget to use the `arg_factory.apply_defaults` decorator.  We mitigate this drawback by making sure that the ArgFactory object generates a helpful error message whenever it is used in an unexpected context.

PiperOrigin-RevId: 505694534