Fix docs-build warnings.
Previously, there were warnings introduced due to dependencies not working
out properly. These warnings would cause the automated build to fail on
readthedocs.io. This change fixes up the versioning requirements so that the
docs builds with no warnings, and updates the documentation to encourage the
`-W` flag which turns warnings into errors when building locally.