Make Tags `cloudpickle`-able.
[Cloudpickle](https://github.com/cloudpipe/cloudpickle) is an alternative to
the standard `pickle` Python module. A key advantage is that it allows
serialization of functions and classes _by value_ instead of _by reference_.
This is important when shipping lambda functions across the network.
Normally, this is totally fine, however when `cloudpickle` `load`'s a
subclass of `fdl.Tag`, the error checking (e.g. to ensure documentation is
present on a tag) fires in `TypeTag.__init__` because the supplied `dct` is
empty. (It's normally set via mutation after instantiating the type object.)

I investigated multiple avenues to customize pickling of Tags, including
`__getstate__`/`__setstate__` and `__reduce__`. However, the custom
`reducer_override` defined on `CloudPickler` doesn't let us hook into the
serialization. Further, you can't customize the reducer via the dispatch
table either.

An alternative I considered (but rejected) was to define a custom subclass of
`Cloudpickler` that provided a custom `reducer_override`.

Cloudpickle is smart enough to realize that if it deserializes a type that
is already available to drop the newly deserialized copy and just use the
existing type. (See the `test_tag_cloudpickle_*` tests that verify object
identity of the deserialized `Tag`s.)

PiperOrigin-RevId: 515533634