absl_flags: Error message should include available names.
Peviously, the error message included the full dotted name that could not be
imported, but it did not indicate which part of the path caused issues, nor
did it include the set of available names (e.g. to help people spot typos or
missing dependencies).

This change augments the error path in `_import_dotted_name` to include more
information in the error message. This change preserves the prefix of the
error messages (as verified by the test suite), but adds an additional
suffix.

PiperOrigin-RevId: 516455809