Improve the Fiddle `absl_flags` example.
This change adds a demonstration for how to stack multiple experiments
or feature launches on top of each other, as well as expands docstrings to
look closer to how they likely should look "in practice".

Note: in order to make Python type-checking work when using Tags in an
`auto_config` function, this change also relaxes the type returned from
`MyTag.new` to be `Any`.
PiperOrigin-RevId: 518166197