Extend signatures.get_type_hints to handle functors.
Previously, dataclass instances with a __call__ method were returning the annotations corresponding to the class, not the __call__ method.

Additionally, rewrites some of the logic to try to more explicitly define the types we expect, and adds a (still internal) top-level module to forward the main functions provided by the signatures module.

PiperOrigin-RevId: 520165799