Remove empty tag sets when flattening a buildable.
While this is a trivial change from the perspective of what's being
serialized, this is quite important due to the use of `__flatten__` powering
the diffing infrastructure. In particular, if a config has tag added, and
subsequently removed, it will now have an empty set associated with that
argument. Before this change, the diffing tooling will find that this
added-and-removed config is different than the one that never had the tag
set in the first place. This is naturally extremely confusing and also a
bug.

PiperOrigin-RevId: 522189292