Unwrap `TaggedValueCls` to preserve the uniform access principle.
Previously, if you had assigned the output of `MyTag.new(some_value)` to a
Fiddle field (e.g. `config.foo.bar`, you now need to always remember to
access the `bar` property as `config.foo.bar.value`. This is relatively
fragile, especially in the presence of Fiddlers that attempt to be generic
across a number of "base" configurations.

This change "unwraps" the `TaggedValueCls` to store the tags in the
`Buildable.__argument_tags__` dictionary, and directly stores the value (if
set) in the `Buildable.__arguments__` dictionary.

There is one "regression" in functionality with this change where if a user
set `TaggedValue` with no value to a `**kwarg` parameter, Fiddle would
prevent `build` from completing with an error about a `TaggedValue` not
having a value. With this change, it's possible to assign a tag to a field
that doesn't have a value associated with it, but then for the object to be
built by Fiddle. An earlier version of this change added such extra checks
to `Config.__build__` to catch this issue, but after discussions within a
subset of the core team, we convinced ourselves that while this _likely_
indicates a bug in user's configuration, it doesn't _always_ indicate one.
So instead, a subsequent change will include a "validation" pass that can
look for these kinds of issues that users can optionally run.

PiperOrigin-RevId: 523450284