Introduce special overrides module for import aliases for modules and symbols.
Refactoring internal fiddle code today, can break config serialization and codegen if the internal refactoring moves around or renames modules and/or symbols. Codegen today solves this by having a special override map in `import_manager`, but there is a similar usecase for config serialization that is unaddressed. This commit introduces a new module that serves as a single store for handling special overrides for codegen, serialization & maybe other usecases in the future.

PiperOrigin-RevId: 552914636