Fixes signature inference for Generic subclasses.
This came up in the context of a user class that was subclassing Protocol... it turns out Protocol subclasses are also Generic subclasses, and this was causing our signature inference to use the signature of __init__ instead of the overall class signature. As a result, when passing arguments positionally, they were incorrectly associated with `self` instead of the correct positional argument, causing duplicate argument errors during `build()`.

I've removed the line switching signature inference to use __init__ for Generics... this doesn't seem like correct behavior to me in general, and as far as I can tell is not covered by any tests. If there is a specific case this was intended to handle, we should revisit and add support for that in another way that doesn't break signature inference for Protocol subclasses.

PiperOrigin-RevId: 590624313