Allow arbitrary keyword argument names with fiddle.Config.
Currently, attempting to build a Config with a "self"
argument fails, because "self" conflicts with arguments
on Buildable.__init__ and Buildable.__build__.

We can avoid the possibility of name conflicts by making
use of positional-only arguments (Python 3.8+). This is
convenient especially for enabling the use of unbound
methods in configs, which can be a nice alternative to
porting existing APIs to use functions.

PiperOrigin-RevId: 674372956