Implement fiddle.absl_flags.DEFINE_fiddle_sweep, a command-line multi-flag like DEFINE_fiddle_config but which supports specifying a sweep of potentially multiple configs to launch.
Sweeps are specified via sweep functions which return dicts of overrides to apply. The overrides can apply either to the arguments to the config: function, or attributes of the returned config (similarly to set: commands).

If multiple sweeps are specified, a product of the sweeps is taken.

Example usage:
my_binary \
    --config 'config:model_config(dropout_rate=0.2)' \
    --config 'sweep:hidden_size_sweep' \
    --config 'sweep:num_layers_sweep(max_layers=10)' \
    --config 'set:layer_norm=True' \
    --config 'fiddler:add_debug_logging' \

PiperOrigin-RevId: 743103489