Make the error message when an unrecognized attribute is used in a Buildable more useful.
We often deal with configs with 1k+ lines and this common error message puts the missing attribute at the top before printing out the entire Buildable. At best, it is difficult to figure out the missing attribute, but usually, it is beyond the pagination limit. This change explicitly lists the Buildable's qualified name and missing attribute at the end so it is easy to identify.

PiperOrigin-RevId: 819920081